pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-extractor"
version = "0.1.0"
description = "Production-ready RAG system for PDF question-answering"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.25.0,<2.0",
    "chromadb>=1.0.0,<2.0",
    "sentence-transformers>=3.4.0,<6.0",
    "ollama>=0.3.0,<1.0",
    "pydantic>=2.6.0,<3.0",
    "pydantic-settings>=2.2.0,<3.0",
    "opentelemetry-api>=1.24.0,<2.0",
    "opentelemetry-sdk>=1.24.0,<2.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0,<2.0",
    "arize-phoenix-otel>=0.6.0,<1.0",
    "click>=8.1.0,<9.0",
    "chainlit>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0",
    "pytest-cov>=5.0.0,<7.0",
    "pytest-mock>=3.14.0,<4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
pdf-extractor = "pdf_extractor.ui.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=pdf_extractor --cov-report=term-missing --cov-fail-under=85"

[tool.coverage.run]
source = ["src/pdf_extractor"]
omit = ["*/ui/cli.py", "*/ui/chainlit_app.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
